Extract CollectCashViewHelper to own file
